requirements.txt
python-dotenv
fastapi[standard]
uvicorn
requests